+++ scilab/modules/mpi/sci_gateway/c/sci_mpi_wait.c
@@ -12,7 +12,6 @@
  * along with this program.
  *
  */
-#include <stdio.h>
 #include <mpi.h>
 #include "api_scilab.h"
 #include "gw_mpi.h"
@@ -22,6 +21,7 @@
 #include "sci_malloc.h"
 #include "serialization.h"
 #include "deserialization.h"
+#include "getOptionalComm.h"
 
 int sci_mpi_wait(char *fname, void* pvApiCtx)
 {
@@ -34,8 +34,25 @@ int sci_mpi_wait(char *fname, void* pvApiCtx)
     int iRequestID;
     double dblRequestID;
 
-    CheckInputArgument(pvApiCtx, 1, 1);
+    CheckInputArgument(pvApiCtx, 1, 2);
     CheckOutputArgument(pvApiCtx, 1, 1);
+
+    // if no optional "comm" is given, return MPI_COMM_WORLD
+    MPI_Comm comm = getOptionalComm(pvApiCtx);
+    if (comm == NULL)
+    {
+        Scierror(999, _("%s: Wrong type for input argument #%s: An MPI communicator expected.\n"), fname, "comm");
+        return 0;
+    }
+
+    if (comm == MPI_COMM_NULL)
+    {
+        // return empty matrix
+        createMatrixOfDouble(pvApiCtx, nbInputArgument(pvApiCtx) + 1, 0, 0, NULL);
+        AssignOutputVariable(pvApiCtx, 1) = nbInputArgument(pvApiCtx) + 1;
+        ReturnArguments(pvApiCtx);
+        return 0;
+    }
 
     sciErr = getVarAddressFromPosition(pvApiCtx, 1, &piAddr);
     if (sciErr.iErr)